Ported OpenIddictMessage tests
